crm_pr/requirements.txt
django-phonenumber-field==5.2.0
phonenumberslite==8.12.36
pytest==6.2.5
django-google-charts==0.1.1